Creates ODriveUART.h and ODriveUART.cpp
